requirements.txt
google-generativeai
dspy-ai>=2.5.0
pydantic>=2.7
typer>=0.12
rich>=13.7
numpy>=1.24
scikit-learn>=1.3
python-dotenv>=1.0
